+++ test/utils/media-type.spec.js
@@ -1,7 +1,7 @@
-import MediaType from "../src/media-type"
+import MediaType from "../../src/utils/media-type"
 
 describe("MediaType", () => {
-  describe("isHumanReadable", () => {
+  describe("#isHumanReadable", () => {
     it("returns true when the content-type is human readable and there's no content-encoding", () => {
       const res = {
         headers: {
@@ -11,6 +11,16 @@ describe("MediaType", () => {
 
       const mediaType = new MediaType(res)
       expect(mediaType.isHumanReadable()).to.be.true
+    })
+
+    it("returns false when content-type is not present", () => {
+      const res = {
+        headers: {
+        }
+      }
+
+      const mediaType = new MediaType(res)
+      expect(mediaType.isHumanReadable()).to.be.false
     })
 
     it("returns false when the content-type is not human readable", () => {
@@ -46,6 +56,40 @@ describe("MediaType", () => {
 
       const mediaType = new MediaType(res)
       expect(mediaType.isHumanReadable()).to.be.false
+    })
+  })
+
+  describe("#isJSON", () => {
+    it("retunrs true when content-type is JSON", () => {
+      const res = {
+        headers: {
+          "content-type": ["application/json"]
+        }
+      }
+
+      const mediaType = new MediaType(res)
+      expect(mediaType.isJSON()).to.be.true
+    })
+
+    it("rerturns false when content-type is not JSON", () => {
+      const res = {
+        headers: {
+          "content-type": ["text/html"]
+        }
+      }
+
+      const mediaType = new MediaType(res)
+      expect(mediaType.isJSON()).to.be.false
+    })
+
+    it("returns false when content-type is not set", () => {
+      const res = {
+        headers: {
+        }
+      }
+
+      const mediaType = new MediaType(res)
+      expect(mediaType.isJSON()).to.be.false
     })
   })
 })