+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.9.1
+- Load tapes from deep directories
+
 ## v1.9.0
 - `responseDecorator` is now called for both matched tapes as well as the initial response returned by the proxied server